requirements.txt
pandas
pyarrow
numpy
Pillow
filetype

# ATS Resume Checker Dependencies
sentence-transformers>=2.2.0
spacy>=3.7.0
scikit-learn>=1.3.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.0.0
textract>=1.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Flask Web App Dependencies
flask==3.0.0
PyPDF2==3.0.1
werkzeug==3.0.1

# Database Dependencies for Production
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
Flask-Migrate>=4.0.0

# Environment and Configuration
python-dotenv>=1.0.0
gunicorn>=21.0.0

# Security and Validation
Flask-Limiter>=3.5.0
Flask-WTF>=1.2.1
Flask-Talisman>=1.1.0
redis>=5.0.0
celery>=5.3.0
prometheus-flask-exporter>=0.23.0
prometheus-client>=0.19.0
marshmallow>=3.20.0
email-validator>=2.1.0
bcrypt>=4.0.0
psutil>=5.9.0